Move ffglib.h to inc
