+++ fontforge/Makefile.am
@@ -104,7 +104,7 @@ pkginclude_HEADERS = autowidth2.h configure-fontforge.h fontforge.h		\
 	uiinterface.h bezctx_ff.h edgelist.h groups.h nonlineartrans.h		\
 	sd.h unicoderange.h bitmapcontrol.h encoding.h ofl.h search.h		\
 	usermenu.h fffreetype.h PfEd.h sfd1.h ffpython.h			\
-	plugins.h sflayoutP.h print.h splinefont.h cvruler.h ffglib.h           \
+	plugins.h sflayoutP.h print.h splinefont.h cvruler.h           \
 	flaglist.h \
 	glif_name_hash.h \
 	glyphcomp.h \
