+++ fontforge/autotrace.c
@@ -52,7 +52,7 @@
 #include <errno.h>		/* for errors */
 #include <dirent.h>		/* for opendir,etc. */
 
-#include "ffglib.h"
+#include <ffglib.h>
 
 int preferpotrace = false;
