+++ fontforge/fontforgeui.h
@@ -28,7 +28,7 @@
 #define _PFAEDITUI_H_
 
 #include <fontforge-config.h>
-#include "ffglib.h"
+#include <ffglib.h>
 #include "fontforgevw.h"
 #include <gprogress.h>
