+++ fontforgeexe/displayfonts.c
@@ -24,7 +24,7 @@
  * OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF
  * ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
  */
-#include "ffglib.h"
+#include <ffglib.h>
 
 #include "autotrace.h"
 #include "cvundoes.h"
