+++ gutils/unicodelibinfo.c
@@ -33,7 +33,7 @@
 
 #include "unicodelibinfo.h"
 #include <ustring.h>
-#include "../fontforge/ffglib.h"
+#include <ffglib.h>
 #include <glib/gprintf.h>
 #include "xvasprintf.h"
