+++ inc/ffglib.h
